+++ docs/release-notes.md
@@ -5,11 +5,12 @@
 
 ### Added
 
-- It is now possible to use connect(), disconnect() instead of a async contextmanager in multithread scenarios.
+- It is now possible to use connect(), disconnect() instead of a async contextmanager in multi-loop calls (multithreading).
 
 ### Fixed
 
 - Database calls are forwarded to subdatabase when possible. This unbreaks using not the returned database object.
+- `force_rollback` works also in multi-loop call (multithreading).
 
 ## 0.9.6